Rename namespace Altinn.Process.Workflow to .Elements
