+++ src/Altinn.Process/BpmnReader.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Xml.Serialization;
 
-using Altinn.Process.Workflow;
+using Altinn.Process.Elements;
 
 namespace Altinn.Process
 {
